+++ src/controllers/index.ts
@@ -5,3 +5,4 @@
 
 export * as authController from './auth.controller.js';
 export * as googleAuthController from './googleAuth.controller.js';
+export * as githubAuthController from './githubAuth.controller.js';